backend/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
passlib[bcrypt]
pydantic[email]
python-jose